requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
yfinance>=0.2.36
pandas>=2.2.0
hmmlearn>=0.3.2
transformers>=4.41.0
accelerate>=0.34.0
scipy>=1.12.0
groq>=0.9.0
python-dotenv
git+https://github.com/amazon-science/chronos-forecasting.git
